moves functions from confusing certificates package to dynakube_props package
