+++ pkg/api/v1beta3/dynakube/otelc_props_test.go
@@ -1,31 +1,30 @@
-package certificates
+package dynakube
 
 import (
 	"testing"
 
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta3/dynakube"
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta3/dynakube/activegate"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestIsAGCertificateNeeded(t *testing.T) {
 	t.Run("remote AG and no trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{},
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{},
 		}
-		assert.False(t, IsAGCertificateNeeded(dk))
+		assert.False(t, dk.IsAGCertificateNeeded())
 	})
 	t.Run("remote AG and trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				TrustedCAs: "test",
 			},
 		}
-		assert.False(t, IsAGCertificateNeeded(dk))
+		assert.False(t, dk.IsAGCertificateNeeded())
 	})
 	t.Run("in-cluster AG and no trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				ActiveGate: activegate.Spec{
 					TlsSecretName: "test-ag-cert",
 					Capabilities: []activegate.CapabilityDisplayName{
@@ -34,11 +33,11 @@ func TestIsAGCertificateNeeded(t *testing.T) {
 				},
 			},
 		}
-		assert.True(t, IsAGCertificateNeeded(dk))
+		assert.True(t, dk.IsAGCertificateNeeded())
 	})
 	t.Run("in-cluster AG and trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				ActiveGate: activegate.Spec{
 					TlsSecretName: "test-ag-cert",
 					Capabilities: []activegate.CapabilityDisplayName{
@@ -48,28 +47,28 @@ func TestIsAGCertificateNeeded(t *testing.T) {
 				TrustedCAs: "test",
 			},
 		}
-		assert.True(t, IsAGCertificateNeeded(dk))
+		assert.True(t, dk.IsAGCertificateNeeded())
 	})
 }
 
 func TestIsCACertificateNeeded(t *testing.T) {
 	t.Run("remote AG and no trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{},
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{},
 		}
-		assert.False(t, IsCACertificateNeeded(dk))
+		assert.False(t, dk.IsCACertificateNeeded())
 	})
 	t.Run("remote AG and trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				TrustedCAs: "test",
 			},
 		}
-		assert.True(t, IsCACertificateNeeded(dk))
+		assert.True(t, dk.IsCACertificateNeeded())
 	})
 	t.Run("in-cluster AG and no trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				ActiveGate: activegate.Spec{
 					TlsSecretName: "test-ag-cert",
 					Capabilities: []activegate.CapabilityDisplayName{
@@ -78,11 +77,11 @@ func TestIsCACertificateNeeded(t *testing.T) {
 				},
 			},
 		}
-		assert.False(t, IsCACertificateNeeded(dk))
+		assert.False(t, dk.IsCACertificateNeeded())
 	})
 	t.Run("in-cluster AG and trustedCAs", func(t *testing.T) {
-		dk := &dynakube.DynaKube{
-			Spec: dynakube.DynaKubeSpec{
+		dk := &DynaKube{
+			Spec: DynaKubeSpec{
 				ActiveGate: activegate.Spec{
 					TlsSecretName: "test-ag-cert",
 					Capabilities: []activegate.CapabilityDisplayName{
@@ -92,6 +91,6 @@ func TestIsCACertificateNeeded(t *testing.T) {
 				TrustedCAs: "test",
 			},
 		}
-		assert.False(t, IsCACertificateNeeded(dk))
+		assert.False(t, dk.IsCACertificateNeeded())
 	})
 }
